requirements.txt
Cython
flask
discord.py
requests
numpy>=1.19.0
pandas>=1.0.5
python-dotenv
aiohttp
aiofiles
base64
re
